made all the getColor routines accept float values
git-svn-id: svn://orca.physics.unc.edu/Orca@10569 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Experiments/Katrin/KatrinDetectorView.m
@@ -99,7 +99,7 @@
 	[super drawRect:rect];
 }
 
-- (NSColor*) getColorForSet:(int)setIndex value:(unsigned long)aValue
+- (NSColor*) getColorForSet:(int)setIndex value:(float)aValue
 {
 	if(setIndex==0)return [focalPlaneColorScale getColorForValue:aValue];
 	else return [vetoColorScale getColorForValue:aValue];
